Refactoring some code.
Program now saves two version of the masks. One with the openCV format to be used in GrabCut. The other is a binary mask with 0 for background and 1 for foreground.

Now the program's input is a configuration file that receives: source directory, mask directory, binary mask directory, file format, index begin, index end.